Link `libunwind` properly on macos (#1093)
On macOS, we need to look for `libunwind` as an LLVM library, rather
than a system-installed dependency. This means adding some additional
logic to the logic that assembles compiler flags when invoking `clang`;
we can't directly look for `@BREW_PREFIX@/opt/llvm/...` because we don't
necessarily know what formula version the user has installed.