llvm-kompile --profile-matching (#1095)
We add a new flag to `llvm-kompile` which, when enabled, instruments the
`k_step` function and the optimized `step_<ordinal>` functions with
calls to a new translation unit, clock.cpp, which contains some very
basic use of `clock_gettime` in order to determine how much time is
being spent doing pattern matching in order to try to apply each rewrite
step. The resulting file contains data about the ordinal of the rule
that applied and the time it took to determine that that rule should
apply. It emits this data on stderr in a tabular format that can be read
easily by a computer. This information is useful in determining how to
optimize pattern matching in cases where it is still not clear where
time is being spent even after `kompile -O3 --iterated-threshold 1` is
passed. When the flag is not passed to the code generator, no
instrumentation is emitted, meaning this is zero-overhead unless it is
enabled.