delete function body before creating eval_ functions
This is needed because previously we handled these with weak symbols,
but the linker can no longer do this for us because everything is now in
the same translation unit. Therefore, we reset these functions so that
they no longer contain their "default" implementations when they get
created by the compiler.