remove remaining calls to getUnqual with type
On llvm 15, the llvm::PointerType::getUnqual(llvm::Type *) function
still creates typed pointers. We don't want to use typed pointers ata
ll, so we remove these call sites and replace them with the getUnqual
overload that uses opaque pointers.

Some further refactoring to condense and clean up the code is probably a
good idea in the future; this makes a lot of the code kind of weird
because I did a purely mechanical translation rather than attempting to
make the code look the way it would if we had written it for the first
time with opaque pointers.