rewrite create_malloc to not call CreateMalloc
On llvm 15, CreateMalloc seems to be broken in some cases when working
with opaque pointers. This is probably due to an llvm bug. However, this
function has a lot of complexity when really it's doing something pretty
simple. So I have replaced the complex version of the code with a fairly
simple version that just creates the desired CallInst directly.