Finalize removal of typed pointers (#1103)
Previously we still had some references to typed pointers in .ll files
and in API calls that still created typed pointers when working with
older versions of LLVM. This led to some invalid bitcode being generated
in ways that wasn't necessarily a problem, but which caused problems
when I was making changes to that code. This PR is a preliminary to
further changes which removes all remaining references to the API
constructor for typed pointers and removes all typed pointers from flat
bitcode files.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>