Refactor `k-rule-apply` tool (#1110)
This PR is a preparatory refactoring that will enable me to fix a
problematic bug in the Kasmer project; I've factored out the changes
here because they include other related cleanups not directly relevant
to the bug I'm fixing in the subsequent PR.

Currently, the `k-rule-apply` tool is explicitly linked against the KORE
AST library in order for it to be able to parse a KORE definition and
search for an axiom with a particular label. This works fine, but on
changing the way that the LLVM backend is linked on macOS, this AST
linkage caused a subtle bug to manifest because of the way that
`k-rule-apply` uses `dlopen` to inspect a shared library compiled by the
backend.[^1]

The fix here is to remove the explicit dependency of this tool on the
AST + Parser libraries, and instead push all the required behaviour of
the tool into the C bindings so that it can be used fully with `dlopen`.

The changes in this PR are as follows:
- Reorganise the existing code in the `k-rule-apply` tool to be more
hygienic in its usage of header files etc.
- Push calls to the KORE parser library and references to the KORE AST
structures out in favour of C-linkage calls to the bindings module,
which is available and compiled into the same shared library as we're
already `dlopen`ing.
- Extract the one non-trivial usage of the AST library (mapping rule
labels to matching function names via their axiom) into the core C
bindings.
- Remove the dependency of this tool on the AST and Parser libraries,
finally enabling the downstream bugfix that motivated this change in the
first place.

[^1]: I have a full writeup of this on the way; in the context of this
PR the specifics aren't important.