Make some code easier to inline (rebased) (#1105)
The llvm backend was not being particularly intelligent about the
ability to inline certain function calls. Here we move all the `.ll`
files in the `runtime` directory into the module used to initialize code
generation. This allows these calls to live in one translation unit and
be better optimized, while also simplifying llvm-kompile-clang.