Better inlining for alloc.cpp (#1121)
This PR improves the inlining for allocations considerably by ensuring
that the bump allocator gets inlined into the code generator. We do this
by means of moving `kore_arena_alloc` to the header file and making it
an inline definition, as well as modifying the build process so that the
bitcode file generated by alloc.cpp gets linked using LLVMs bitcode
linker into the code generator module prior to optimization.