Updating `StringPattern` to return its contents as Python Bytes (#1126)
This PR modifies the behavior of `StringPattern` binding. It returns the
content of `kore_string_pattern` as `py::bytes`.

The `value` must be encoded with `raw_unicode_scape` in order to get
`assert StringPattern(value).contents == value`.

We fixed a test_pattern to reflect this behavior and added 2 new tests
that show the features we're enabling with this modification, an initial
Unicode support, among them.