Buffered reads/writes for hint generation with shared memory (#1127)
This PR adds implementation for buffered reads/writes from and to the
shared memory ringbuffer during hint generation. The previous
implementation would synchronize reads and writes at the granularity of
one byte which is not ideal for performance. In this PR, we add support
for synchronization across reads and writes of larger sizes. Both the
ringbuffer-based proof trace reader and writer classes are updated to
buffer incoming reads and writes: They collect data into an internal
buffer and only interacting with the shared memory ringbuffer when that
buffer is full in case of the writer or empty in case of the reader.