Abstract proof trace writer at event level (#1136)
This PR refactors the proof trace writer to be abstract at the event
generation level, rather than the byte stream level. We do this in order
to be able to use alternative ways of communicating the proof hints that
do not require writing the trace to a file, such as triggering of
callback invocations.

The PR can be reviewed commit by commit:
- the 1st commit removes an unused flag from main.ll
- the 2nd commit adds the new abstract `proof_trace_writer` and a
subclass that writes in a file as before.
- the 3rd commit removes the old `proof_trace_writer` that operated at
the level of writing byte streams.