Un-nest proof hint events for arguments of functions/hooks (#1140)
The events that correspond to evaluation of arguments of functions and
hooks are currently nested inside the corresponding function/hook event.
There is no need for that and it creates various complications in the
processing of the proof hint trace. For this reason, this PR un-nests
those events so that they appear before their corresponding
function/hook event.