Add pattern matching support for list random access patterns (#1143)
This is not rigorously tested, nor is it likely to be in the near future
because the code it was written to support ended up not being worth
merging. However, it is broadly useful code that may find value in the
future, so I would like to get it merged.

It also includes a few fixes to python scripts in the repo that had gone
slightly stale.