Only emitting side_conditions if all occurrences of the same variable are equal (#1155)
In this PR, we inverse the order in which we create the decision tree
when a rule has a side condition. Previously, even when we had 2 or more
variables with the same name having different values, the side condition
decision trees were generated, and we only evaluated the variables
matching after evaluating the side condition.

After this PR, we inverse this order and evaluate the variables matching
first and then the side condition.
This issue was discovered in the context of side conditions proof events
being emitted and being true but without ever being applied, as we have
a mismatch in the variable names and values.

This PR fixes: https://github.com/Pi-Squared-Inc/pi2/issues/2124

We added a regression test that reproduced this issue, and it's now
fixed!