Minor optimization to search (#1170)
In the final version of the ULM, we want to enforce that semantics are
deterministic. We do this using the search feature and the
`--execute-to-branch` option which treats all states with more than one
successor as final states and does not rewrite them further.

I tested this by running the ethereum test suite with this option and
all the tests passed (with a few trivial changes to the semantics). A
minor change resulted from this effort: if only a single state is active
at a given time, we do not need to do the expensive equality check
performed by `erase`. It is sufficient to simply clear the hash set.